MAINT: Move lazydict module out of io directory
