+++ doc/source/dev_reference/lazydict.rst
@@ -1,0 +1,7 @@
+==============
+pyart.lazydict
+==============
+
+Lazy load dictionary.
+
+.. automodule:: pyart.lazydict
