+++ pyart/io/mdv_radar.py
@@ -17,7 +17,7 @@ from netCDF4 import date2num
 from ..config import FileMetadata, get_fillvalue
 from ..core.radar import Radar
 from .common import make_time_unit_str, _test_arguments, prepare_for_read
-from .lazydict import LazyLoadDict
+from ..lazydict import LazyLoadDict
 from . import mdv_common
