+++ pyart/io/nexrad_archive.py
@@ -23,7 +23,7 @@ from ..config import FileMetadata, get_fillvalue
 from ..core.radar import Radar
 from .common import make_time_unit_str, _test_arguments, prepare_for_read
 from .nexrad_level2 import NEXRADLevel2File
-from .lazydict import LazyLoadDict
+from ..lazydict import LazyLoadDict
 
 
 def read_nexrad_archive(filename, field_names=None, additional_metadata=None,
